+++ Runtime/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Runtime.CompilerServices;
 using Unity.Cloud.Common;
 
-[assembly: ApiSourceVersion("com.unity.cloud.identity", "0.15.0")]
+[assembly: ApiSourceVersion("com.unity.cloud.identity", "0.15.2")]
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Tests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // to allow moq implementations for internal interfaces