+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.15.2] - 2023-04-27
+
+### Added
+- Added a documentation section on `Managed Stripping Level` for build settings.
+
+### Fixed
+- Reload browser page issue on WebGL.
+
+## [0.15.1] - 2023-04-13
+
+### Added
+- `SignOutUrl` property in `PkceConfiguration`.
+- `IAuthenticationPlatformSupport.GetRedirectUri` method accepts an optional string parameter to support any redirection route.
+
+### Changed
+- `LogoutAsync` method in `IUrlRedirectionAuthenticator` interface and its derived implementations accepts an optional boolean value to clear the browser cache.
+
 ## [0.15.0] - 2023-03-30
 
 ### Added